feat(attendanceRequests): isolate controllers
